% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelall.R
\docType{data}
\name{modelall}
\alias{modelall}
\title{All data for modeling}
\format{
A data with 1089 rows and 199325 variables
}
\source{
GSE126250
}
\usage{
modelall
}
\description{
THis data is used for modeling.
modelall
}
\keyword{datasets}
