% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test}
\alias{test}
\title{All data for test}
\format{
A data with 100 rows and 25482 variables
}
\source{
GSE126250
}
\usage{
test
}
\description{
THis data is for test
test
}
\keyword{datasets}
