% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{TH2A_probability}
\alias{TH2A_probability}
\title{predict of the TH2A prob}
\usage{
TH2A_probability(
  xtest,
  Panelgene = c("CACNA1D", "IL1RL1", "IL17RB", "IL5", "INPP1", "HPGDS", "EGLN3")
)
}
\arguments{
\item{xtest}{input dataframe of your scRNA data.}

\item{Panelgene}{the panel gene you decided.}
}
\value{
a new col of your data with prediction either TH2A (return 1) or Non-TH2A (return 0).
}
\description{
predict of the TH2A prob
}
\examples{
TH2A_probability(test)
}
