% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the document of the package.R
\docType{package}
\name{TH2Aphenotyping}
\alias{TH2Aphenotyping}
\title{TH2Aphenotyping}
\description{
This package is aim to identified the TH2A cells based on the single-cell sequencing data.
}
\section{TH2A}{
NA
}

